# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.BucketType:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Category:
    properties:
      bucket_type:
        allOf:
        - $ref: '#/definitions/models.BucketType'
        description: RELATIONSHIPS
      bucket_type_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      type:
        enum:
        - expense
        - income
        type: string
      updated_at:
        type: string
    required:
    - name
    - type
    type: object
  models.CategoryRequest:
    properties:
      bucket_type_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      type:
        enum:
        - expense
        - income
        type: string
    required:
    - bucket_type_id
    - name
    - type
    type: object
...
  models.DeleteCategoryResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      type:
        enum:
        - expense
        - income
        type: string
      updated_at:
        type: string
    required:
    - name
    - type
    type: object
...
  utils.PaginationResponse-array_models_Category:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  utils.Response-any:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
...
  utils.Response-array_models_Category:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      error:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
...
  utils.Response-models_Category:
    properties:
      data:
        $ref: '#/definitions/models.Category'
      error:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
...
  utils.Response-models_DeleteCategoryResponse:
    properties:
      data:
        $ref: '#/definitions/models.DeleteCategoryResponse'
      error:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
...
  utils.ResponseWithPagination-array_models_Category:
    properties:
      data:
        $ref: '#/definitions/utils.PaginationResponse-array_models_Category'
      error:
        type: boolean
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: gonanggoneng@gmail.com
    name: Anang
    url: http://github.com/anangkf
  description: API documentation for Expense Tracker project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get all categories for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: id
        description: Sort by field
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Sort order (asc or desc)
        in: query
        name: order
        type: string
      - description: Filter by category name
        in: query
        name: name
        type: string
      - description: Filter by category type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseWithPagination-array_models_Category'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Get categories by user ID
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category for the authenticated user
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response-models_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category for the authenticated user
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-models_DeleteCategoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get a category by ID for the authenticated user
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-models_Category'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Get a category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update a category for the authenticated user
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-models_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /categories/default:
    get:
      consumes:
      - application/json
      description: Get default categories for all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-array_models_Category'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Get default categories
      tags:
      - categories
  /categories/multiple:
    post:
      consumes:
      - application/json
      description: Create multiple categories for the authenticated user
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/models.CategoryRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response-array_models_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Create multiple categories
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"